refactor: new ui path
